feat(settings): improve mobile profile view navigation and reorder dropdown items
