+++ app/components/MobileDropdown.tsx
@@ -40,7 +40,7 @@ import { TransferForm } from "./TransferForm";
 import { EarnWalletForm } from "./EarnWalletForm";
 import { EarnConsentModal } from "./EarnConsentModal";
 import { CopyAddressWarningModal } from "./CopyAddressWarningModal";
-import ProfileDrawer from "./ProfileDrawer";
+import ProfileView from "./ProfileView";
 import WalletMigrationModal from "./WalletMigrationModal";
 import { useEarnAccess } from "../hooks/useEarnAccess";
 import { isEarnUiVisible } from "../lib/earnFeature";
@@ -66,7 +66,6 @@ export const MobileDropdown = ({
   const [isLoggingOut, setIsLoggingOut] = useState(false);
   const [isWarningModalOpen, setIsWarningModalOpen] = useState(false);
   const [isMigrationModalOpen, setIsMigrationModalOpen] = useState(false);
-  const [isProfileDrawerOpen, setIsProfileDrawerOpen] = useState(false);
 
   const { selectedNetwork, setSelectedNetwork } = useNetwork();
   const { currentStep } = useStep();
@@ -413,9 +412,15 @@ export const MobileDropdown = ({
                           isLoggingOut={isLoggingOut}
                           onBack={() => setCurrentView("wallet")}
                           onOpenProfile={() => {
-                            onClose();
-                            setIsProfileDrawerOpen(true);
+                            setCurrentView("profile");
                           }}
+                        />
+                      )}
+
+                      {currentView === "profile" && (
+                        <ProfileView
+                          layout="sheet"
+                          onBack={() => setCurrentView("settings")}
                         />
                       )}
 
@@ -474,10 +479,6 @@ export const MobileDropdown = ({
         onClose={() => setIsMigrationModalOpen(false)}
       />
 
-      <ProfileDrawer
-        isOpen={isProfileDrawerOpen}
-        onClose={() => setIsProfileDrawerOpen(false)}
-      />
     </>
   );
 };
